shared-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile shared-requirements.in
#
asgiref==3.3.4
    # via django
django==3.2
    # via
    #   -r shared-requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r shared-requirements.in
pytz==2021.1
    # via django
rules==3.0
    # via -r shared-requirements.in
sqlparse==0.4.1
    # via django
